requirements.txt
Django==5.1.7
django-allauth==0.65.4
django-crispy-forms==2.3
django-ckeditor==6.7.2
Pillow==11.1.0
python-dotenv==1.0.1
requests==2.32.3
openai==1.65.5 